Remove 'nwchemex-' prefixes and template documentation
