+++ spack_repo/nwchemex/core/packages/chemist/package.py
@@ -1,36 +1,19 @@
-# Copyright 2013-2024 Lawrence Livermore National Security, LLC and other
-# Spack Project Developers. See the top-level COPYRIGHT file for details.
+# Copyright 2025-2026 NWChemEx Developers.
 #
-# SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
-# ----------------------------------------------------------------------------
-# If you submit this package back to Spack as a pull request,
-# please first remove this boilerplate and all FIXME comments.
-#
-# This is a template package file for Spack.  We've put "FIXME"
-# next to all the things you'll want to change. Once you've handled
-# them, you can save this file and test your package like this:
-#
-#     spack install nwchemex-simde
-#
-# You can edit this file again by typing:
-#
-#     spack edit nwchemex-simde
-#
-# See the Spack documentation for more information on packaging.
-# ----------------------------------------------------------------------------
+# SPDX-License-Identifier: Apache-2.0
 
 from spack import package as pkg
+
 from spack_repo.nwchemex.common.mixins import NWChemExBasePybindings
 
 
-class NwchemexSimde(NWChemExBasePybindings):
+class Chemist(NWChemExBasePybindings):
     """Generic, helpful C++ classes used by the NWChemEx project."""
 
-    project = "SimDE"
+    project = "Chemist"
 
     homepage = f"https://github.com/NWChemEx/{project}"
-    url = f"https://github.com/NWChemEx/{project}/archive/refs/tags/v0.0.53.tar.gz"
+    url = f"https://github.com/NWChemEx/{project}/archive/refs/tags/v1.3.18.tar.gz"
     git = f"https://github.com/NWChemEx/{project}.git"  # For the latest commit
 
     # Versions are hosted under GitHub tags right now
@@ -47,8 +30,8 @@ class NwchemexSimde(NWChemExBasePybindings):
 
     # Versions from git tags
     pkg.version(
-        "0.0.53",
-        sha256="c95b818c0151a38190eebdaf41cc19fe04227ec827aef30293f959751a4b0bed",
+        "1.3.19",
+        sha256="dc1adf754ce9a532ee4b13461aaaf29488a2d6d5f34aaa416a33bd621abb8c29",
     )
 
     pkg.variant(
@@ -58,29 +41,36 @@ class NwchemexSimde(NWChemExBasePybindings):
         sticky=True,
     )
 
+    # Runtime dependencies
+    pkg.depends_on("boost")
+
     # First-party
+    pkg.depends_on("nwchemex-utilities")
     pkg.depends_on(
-        "nwchemex-chemist+python",
+        "nwchemex-parallelzone+python",
         type=("build", "link", "run"),
         when="+python",
     )
     pkg.depends_on(
-        "nwchemex-chemist~python",
-        type=("build", "link", "run"),
-        when="~python",
-    )
-    pkg.depends_on(
-        "nwchemex-pluginplay+python",
-        type=("build", "link", "run"),
-        when="+python",
-    )
-    pkg.depends_on(
-        "nwchemex-pluginplay~python",
+        "nwchemex-parallelzone~python",
         type=("build", "link", "run"),
         when="~python",
     )
 
-    pkg.depends_on("sigma+eigen", when="+sigma")
+    pkg.depends_on(
+        "nwchemex-tensorwrapper+python",
+        type=("build", "link", "run"),
+        when="+python",
+    )
+    pkg.depends_on(
+        "nwchemex-tensorwrapper~python",
+        type=("build", "link", "run"),
+        when="~python",
+    )
+
+    # Although we have a variant, technically it is not a direct dependency
+    # of this package
+    # pkg.depends_on("sigma+eigen", when="+sigma")
 
     # Start with CMaize sanity check locations
     sanity_check_is_dir = NWChemExBasePybindings.cmaize_sanity_check_dirs(
