+++ spack_repo/nwchemex/core/packages/integrals/package.py
@@ -1,36 +1,19 @@
-# Copyright 2013-2024 Lawrence Livermore National Security, LLC and other
-# Spack Project Developers. See the top-level COPYRIGHT file for details.
+# Copyright 2025-2026 NWChemEx Developers.
 #
-# SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
-# ----------------------------------------------------------------------------
-# If you submit this package back to Spack as a pull request,
-# please first remove this boilerplate and all FIXME comments.
-#
-# This is a template package file for Spack.  We've put "FIXME"
-# next to all the things you'll want to change. Once you've handled
-# them, you can save this file and test your package like this:
-#
-#     spack install nwchemex-simde
-#
-# You can edit this file again by typing:
-#
-#     spack edit nwchemex-simde
-#
-# See the Spack documentation for more information on packaging.
-# ----------------------------------------------------------------------------
+# SPDX-License-Identifier: Apache-2.0
 
 from spack import package as pkg
+
 from spack_repo.nwchemex.common.mixins import NWChemExBasePybindings
 
 
-class NwchemexScf(NWChemExBasePybindings):
+class Integrals(NWChemExBasePybindings):
     """Generic, helpful C++ classes used by the NWChemEx project."""
 
-    project = "SCF"
+    project = "Integrals"
 
     homepage = f"https://github.com/NWChemEx/{project}"
-    url = f"https://github.com/NWChemEx/{project}/archive/refs/tags/v0.0.23.tar.gz"
+    url = f"https://github.com/NWChemEx/{project}/archive/refs/tags/v0.0.30.tar.gz"
     git = f"https://github.com/NWChemEx/{project}.git"  # For the latest commit
 
     # Versions are hosted under GitHub tags right now
@@ -47,20 +30,24 @@ class NwchemexScf(NWChemExBasePybindings):
 
     # Versions from git tags
     pkg.version(
-        "0.0.23",
-        sha256="b175c15e8c814cd288817c970f4e049c7eab248975ff7c891d8927d7555d0cd8",
+        "0.0.30",
+        sha256="894bec1a6be2ec28302fde9f26f9b2961b43cfe4d15c037005fd266295dd7f3b",
     )
 
-    # For building GauXC, I think
-    pkg.depends_on("c", type="build")
+    pkg.variant(
+        "sigma",
+        default=False,
+        description="Enable Sigma for uncertainty tracking",
+        sticky=True,
+    )
 
-    # TODO: Create this package
-    # pkg.depends_on("gauxc")
-    pkg.depends_on("eigen")
-    pkg.depends_on("mpi")
-    pkg.depends_on("py-numpy")
-    # Uncomment when GauXC/Libxc interactions are sorted out
-    # pkg.depends_on("libxc")
+    # The "tune" variant is not available prior to v2.6
+    # TODO: A value of "tune=none" or any of the molgw-* options likely break
+    # the unit tests, but I don't know how to add them as conflicts yet.
+    pkg.depends_on("libint@2.6:")
+    # Although we have a variant, technically it is not a direct dependency
+    # of this package
+    # pkg.depends_on("sigma+eigen", when="+sigma")
 
     # First-party
     pkg.depends_on(
@@ -82,3 +69,14 @@ class NwchemexScf(NWChemExBasePybindings):
         project.lower()
     )
     # Append more sanity checks as needed
+
+    def cmake_args(self):
+        args = super().cmake_args()
+
+        args.extend(
+            [
+                self.define_from_variant("ENABLE_SIGMA", "sigma"),
+            ]
+        )
+
+        return args
