+++ spack_repo/nwchemex/core/packages/parallelzone/package.py
@@ -1,30 +1,15 @@
-# Copyright 2013-2024 Lawrence Livermore National Security, LLC and other
-# Spack Project Developers. See the top-level COPYRIGHT file for details.
+# Copyright 2025-2026 NWChemEx Developers.
 #
-# SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
-# ----------------------------------------------------------------------------
-# If you submit this package back to Spack as a pull request,
-# please first remove this boilerplate and all FIXME comments.
-#
-# This is a template package file for Spack.  We've put "FIXME"
-# next to all the things you'll want to change. Once you've handled
-# them, you can save this file and test your package like this:
-#
-#     spack install nwchemex-parallelzone
-#
-# You can edit this file again by typing:
-#
-#     spack edit nwchemex-parallelzone
-#
-# See the Spack documentation for more information on packaging.
-# ----------------------------------------------------------------------------
+# SPDX-License-Identifier: Apache-2.0
 
 from spack.package import depends_on, license, maintainers, variant, version
+
 from spack_repo.nwchemex.common.mixins import NWChemExBasePybindings
 
+# TODO: Use from spack import package as pkg like other packages
 
-class NwchemexParallelzone(NWChemExBasePybindings):
+
+class Parallelzone(NWChemExBasePybindings):
     """Generic, helpful C++ classes used by the NWChemEx project."""
 
     project = "ParallelZone"
@@ -44,6 +29,8 @@ class NwchemexParallelzone(NWChemExBasePybindings):
 
     maintainers("ryanmrichard", "jwaldrop107", "zachcran")
     license("Apache-2.0", checked_by="zachcran")
+
+    version("develop", branch="find_package_wrapper")
 
     # Versions from git tags
     version(
@@ -75,7 +62,10 @@ class NwchemexParallelzone(NWChemExBasePybindings):
     # Runtime dependencies
     depends_on("mpi")
     depends_on("spdlog")
-    depends_on("cereal@1.3.0")  # v1.3.1 changed the installed target...
+    depends_on(
+        "cereal@1.3.1:", when="@0.1.41:"
+    )  # v1.3.1 changed the installed target from "cereal" to "cereal::cereal"
+    depends_on("cereal@:1.3.0", when="@:0.1.40")
     depends_on("papi", when="+papi")
 
     # Start with CMaize sanity check locations
