+++ spack_repo/nwchemex/core/packages/tensorwrapper/package.py
@@ -1,36 +1,19 @@
-# Copyright 2013-2024 Lawrence Livermore National Security, LLC and other
-# Spack Project Developers. See the top-level COPYRIGHT file for details.
+# Copyright 2025-2026 NWChemEx Developers.
 #
-# SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
-# ----------------------------------------------------------------------------
-# If you submit this package back to Spack as a pull request,
-# please first remove this boilerplate and all FIXME comments.
-#
-# This is a template package file for Spack.  We've put "FIXME"
-# next to all the things you'll want to change. Once you've handled
-# them, you can save this file and test your package like this:
-#
-#     spack install nwchemex-simde
-#
-# You can edit this file again by typing:
-#
-#     spack edit nwchemex-simde
-#
-# See the Spack documentation for more information on packaging.
-# ----------------------------------------------------------------------------
+# SPDX-License-Identifier: Apache-2.0
 
 from spack import package as pkg
+
 from spack_repo.nwchemex.common.mixins import NWChemExBasePybindings
 
 
-class NwchemexIntegrals(NWChemExBasePybindings):
+class Tensorwrapper(NWChemExBasePybindings):
     """Generic, helpful C++ classes used by the NWChemEx project."""
 
-    project = "Integrals"
+    project = "TensorWrapper"
 
     homepage = f"https://github.com/NWChemEx/{project}"
-    url = f"https://github.com/NWChemEx/{project}/archive/refs/tags/v0.0.30.tar.gz"
+    url = f"https://github.com/NWChemEx/{project}/archive/refs/tags/v0.0.55.tar.gz"
     git = f"https://github.com/NWChemEx/{project}.git"  # For the latest commit
 
     # Versions are hosted under GitHub tags right now
@@ -47,8 +30,8 @@ class NwchemexIntegrals(NWChemExBasePybindings):
 
     # Versions from git tags
     pkg.version(
-        "0.0.30",
-        sha256="894bec1a6be2ec28302fde9f26f9b2961b43cfe4d15c037005fd266295dd7f3b",
+        "0.0.62",
+        sha256="a526418836e0fff1362d4ff2c9131ae9eb2d509ae53148ac597f8de8444af9ca",
     )
 
     pkg.variant(
@@ -58,25 +41,30 @@ class NwchemexIntegrals(NWChemExBasePybindings):
         sticky=True,
     )
 
-    # The "tune" variant is not available prior to v2.6
-    # TODO: A value of "tune=none" or any of the molgw-* options likely break
-    # the unit tests, but I don't know how to add them as conflicts yet.
-    pkg.depends_on("libint@2.6:")
-    # Although we have a variant, technically it is not a direct dependency
-    # of this package
-    # pkg.depends_on("sigma+eigen", when="+sigma")
+    # Runtime dependencies
+    pkg.depends_on("boost")
+    pkg.depends_on("eigen", type=("build", "link", "run"))
+    pkg.depends_on("py-numpy", when="+python", type=("build", "run"))
 
     # First-party
+    pkg.depends_on("nwchemex-utilities")
     pkg.depends_on(
-        "nwchemex-simde+python",
+        "nwchemex-parallelzone~python",
+        type=("build", "link", "run"),
+        when="~python",
+    )
+    pkg.depends_on(
+        "nwchemex-parallelzone+python",
         type=("build", "link", "run"),
         when="+python",
     )
     pkg.depends_on(
-        "nwchemex-simde~python",
+        "py-numpy",
         type=("build", "link", "run"),
-        when="~python",
+        when="+python",
     )
+
+    pkg.depends_on("sigma+eigen", when="+sigma")
 
     # Start with CMaize sanity check locations
     sanity_check_is_dir = NWChemExBasePybindings.cmaize_sanity_check_dirs(
